Data.Text.Zipper

Description

This module provides a two-dimensional text zipper data structure. This structure represents a body of text and an editing cursor which can be moved throughout the text, along with a set of editing transformations.

Text zippers are generalized over the set of data types that might be used to store lists of characters (e.g., String, Text, etc.). As a result, the most general way to create a text zipper is to use mkZipper and provide all of the functions required to manipulate the underlying text data.

Implementations using Text and String are provided.

Documentation

data TextZipper a

Construction and extraction

mkZipper

Arguments

:: Monoid a

=> (Char -> a)
A singleton constructor.

-> (Int -> a -> a)
drop.

-> (Int -> a -> a)
take.

-> (a -> Int)
length.

-> (a -> Char)
last.

-> (a -> a)
init.

-> (a -> Bool)
null.

-> (a -> [a])
lines.

-> (a -> [Char])
toList.

-> [a]
The initial lines of text.

-> Maybe Int
Limit to this many lines of text (Nothing means no limit).

-> TextZipper a

Create a zipper using a custom text storage type. Takes the initial text as well as all of the functions necessary to manipulate the underlying text values.

textZipper :: [Text] -> Maybe Int -> TextZipper Text

Construct a zipper from Text values.

stringZipper :: [String] -> Maybe Int -> TextZipper String

Construct a zipper from list values.

clearZipper :: Monoid a => TextZipper a -> TextZipper a

Empty a zipper.

vectorZipper :: [Vector Char] -> Maybe Int -> TextZipper (Vector Char)

Construct a zipper from vectors of characters.

getText :: Monoid a => TextZipper a -> [a]

Get the text contents of the zipper.

currentLine :: Monoid a => TextZipper a -> a

The line of text on which the zipper's cursor currently resides.

cursorPosition :: TextZipper a -> (Int, Int)

Get the cursor position of the zipper; returns (row, col). row ranges from [0..num_rows-1] inclusive; col ranges from [0..length of current line] inclusive. Column values equal to line width indicate a cursor that is just past the end of a line of text.

lineLengths :: Monoid a => TextZipper a -> [Int]

Return the lengths of the lines in the zipper.

getLineLimit :: TextZipper a -> Maybe Int

Get the line limit, if any, for a zipper.

Navigation functions

moveCursor :: Monoid a => (Int, Int) -> TextZipper a -> TextZipper a

Move the cursor to the specified row and column. Invalid cursor positions will be ignored. Valid cursor positions range as described for cursorPosition.

moveRight :: Monoid a => TextZipper a -> TextZipper a

Move the cursor right by one position. If the cursor is at the end of a line, the cursor is moved to the first position of the following line (if any).

moveLeft :: Monoid a => TextZipper a -> TextZipper a

Move the cursor left by one position. If the cursor is at the beginning of a line, the cursor is moved to the last position of the preceding line (if any).

moveUp :: Monoid a => TextZipper a -> TextZipper a

Move the cursor up by one row. If there no are rows above the current one, move to the first position of the current row. If the row above is shorter, move to the end of that row.

moveDown :: Monoid a => TextZipper a -> TextZipper a

Move the cursor down by one row. If there are no rows below the current one, move to the last position of the current row. If the row below is shorter, move to the end of that row.

gotoEOL :: Monoid a => TextZipper a -> TextZipper a

Move the cursor to the end of the current line.

gotoBOL :: Monoid a => TextZipper a -> TextZipper a

Move the cursor to the beginning of the current line.

Inspection functions

currentChar :: TextZipper a -> Maybe Char

Get the Char on which the cursor currently resides. If the cursor is at the end of the text or the text is empty return Nothing.

nextChar :: Monoid a => TextZipper a -> Maybe Char

Get the Char after the cursor position. If the cursor is at the end of a line return the first character of the next line, or if that one is empty as well, return Nothing.

previousChar :: Monoid a => TextZipper a -> Maybe Char

Get the Char before the cursor position. If the cursor is at the beginning of the text, return Nothing

Editing functions

insertChar :: Monoid a => Char -> TextZipper a -> TextZipper a

Insert a character at the current cursor position.

If the character is a newline, break the current line.

If the character is non-printable, ignore it.

Otherwise insert the character and move the cursor one position to the right.

insertMany :: Monoid a => a -> TextZipper a -> TextZipper a

Insert many characters at the current cursor position. Move the cursor to the end of the inserted text.

deletePrevChar :: (Eq a, Monoid a) => TextZipper a -> TextZipper a

Delete the character preceding the cursor position, and move the cursor backwards by one character.

deleteChar :: Monoid a => TextZipper a -> TextZipper a

Delete the character at the cursor position. Leaves the cursor position unchanged. If the cursor is at the end of a line of text, this combines the line with the line below.

breakLine :: Monoid a => TextZipper a -> TextZipper a

Insert a line break at the current cursor position.

killToEOL :: Monoid a => TextZipper a -> TextZipper a

Remove all text from the cursor position to the end of the current line. If the cursor is at the beginning of a line and the line is empty, the entire line will be removed.

killToBOL :: Monoid a => TextZipper a -> TextZipper a

Remove all text from the cursor position to the beginning of the current line.

transposeChars :: Monoid a => TextZipper a -> TextZipper a

Transpose the character before the cursor with the one at the cursor position and move the cursor one position to the right. If the cursor is at the end of the current line, transpose the current line's last two characters.